Create new fog.upload script
This commit contains unfinished work on a new fog.upload script and is likely
to be very buggy. At this point only the resizable imaging job code is
complete.

This script attempts to be more OS agnostic. As a result, it may cause issues
with older images. It will likely require work on support functions. Be aware
of this when testing.

Note that functions found in this script may or may not be integrated into the
function library files on a later commit.