Excluding virtual interfaces
In my setup, the machine that I am using to install FOG has two IPs related to one interface (192.168.1.3 for eth0 and 192.168.1.253 for eth0:1). One is the real interface and the second one is a virtual interface. Adding the grep -v $interface. excludes all virtual interfaces created leaving only the real interface resulting in ipaddress=192.168.1.3 instead of ipaddress=192.168.1.3\n192.168.1.254.
Without this correction, the variable ipaddress had both IPs as value leading to a broken Apache configuration file failing Apache's execution.
I am not sure if this is relevant to the project however I decided to post it here and let you guys evaluate if this is useful or not. 
Best regards and many thanks for the outstanding work.